Update simulation to be compatible with grakn 2.0 (#98)
## What is the goal of this PR?

Update simulation to be compatible with grakn 2.0

## What are the changes implemented in this PR?

- Update the syntax and client code to 2.0

+++ common/action/read/ProductsInContinentAction.java
@@ -23,7 +23,7 @@ import grakn.simulation.common.world.World;
 import java.util.ArrayList;
 import java.util.List;
 
-public abstract class ProductsInContinentAction<DB_OPERATION extends DbOperation> extends ReadAction<DB_OPERATION, List<Double>> {
+public abstract class ProductsInContinentAction<DB_OPERATION extends DbOperation> extends ReadAction<DB_OPERATION, List<Long>> {
     protected final World.Continent continent;
 
     public ProductsInContinentAction(DB_OPERATION dbOperation, World.Continent continent) {
